fix: negative values
The problem was that whenever we last saw a "Dash", we consider the next digit to be negative even though there was a space between the two characters.
To solve the problem, I remember if we saw a Dash. If so, and if the current char is not a space, then we know we are in a "-X" situation.
If the current char is a space, then the Dash is used as a binary operator.
Everything works fine now :).